Restructred packages modles.go -> models package PostgresDB struct -> postgresdb package
